+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 ### Features
 - Summary component
 - Line ups component
+
+## Version 0.0.3
+### Features
+- Standings component
+- Match Incidents component